Move Theano-dependent code in the algos package
The following modules were moved into the garage.theano package:
- ddpg
- erwr
- npo
- ppo
- reps
- tnpg
- trpo
- vpg

The corresponding imports were modified as well.
