+++ contrib/alexbeloi/examples/vpgis_cartpole.py
@@ -1,9 +1,9 @@
 from contrib.alexbeloi.is_sampler import ISSampler
-from garage.algos import VPG
 from garage.baselines import LinearFeatureBaseline
 from garage.envs import normalize
 from garage.envs.box2d import CartpoleEnv
 from garage.policies import GaussianMLPPolicy
+from garage.theano.algos import VPG
 """
 Example using VPG with ISSampler, iterations alternate between live and
 importance sampled iterations.
