+++ contrib/rllab_hyperopt/example/task.py
@@ -1,8 +1,8 @@
-from garage.algos import TRPO
 from garage.baselines import LinearFeatureBaseline
 from garage.envs import normalize
 from garage.envs.box2d import CartpoleEnv
 from garage.policies import GaussianMLPPolicy
+from garage.theano.algos import TRPO
 
 
 def run_task(v):
