+++ contrib/ros/envs/sawyer/launchers/trpo_real_sawyer_pnp.py
@@ -2,11 +2,11 @@ import numpy as np
 import rospy
 
 from contrib.ros.envs.sawyer.pick_and_place_env import PickAndPlaceEnv
-from garage.algos.trpo import TRPO
 from garage.baselines.linear_feature_baseline import LinearFeatureBaseline
 from garage.envs.util import spec
 from garage.misc.instrument import run_experiment
 from garage.policies.gaussian_mlp_policy import GaussianMLPPolicy
+from garage.theano.algos import TRPO
 
 INITIAL_ROBOT_JOINT_POS = {
     'right_j0': -0.140923828125,
