+++ examples/theano/ddpg_cartpole.py
@@ -1,10 +1,10 @@
-from garage.algos import DDPG
 from garage.envs import normalize
 from garage.envs.box2d import CartpoleEnv
 from garage.exploration_strategies import OUStrategy
 from garage.misc.instrument import run_experiment
 from garage.policies import DeterministicMLPPolicy
 from garage.q_functions import ContinuousMLPQFunction
+from garage.theano.algos import DDPG
 from garage.theano.envs import TheanoEnv
