+++ examples/theano/trpo_cartpole_pickled.py
@@ -1,9 +1,9 @@
-from garage.algos import TRPO
 from garage.baselines import LinearFeatureBaseline
 from garage.envs import normalize
 from garage.envs.box2d import CartpoleEnv
 from garage.misc.instrument import run_experiment
 from garage.policies import GaussianMLPPolicy
+from garage.theano.algos import TRPO
 from garage.theano.envs import TheanoEnv
