+++ examples/theano/trpo_cartpole_recurrent.py
@@ -1,4 +1,3 @@
-from garage.algos import TRPO
 from garage.baselines import LinearFeatureBaseline
 from garage.envs import normalize
 from garage.envs.box2d import CartpoleEnv
@@ -6,6 +5,7 @@ from garage.misc.instrument import run_experiment
 from garage.optimizers import ConjugateGradientOptimizer
 from garage.optimizers import FiniteDifferenceHvp
 from garage.policies import GaussianGRUPolicy
+from garage.theano.algos import TRPO
 from garage.theano.envs import TheanoEnv
