+++ examples/theano/trpo_sawyer_reacher.py
@@ -1,10 +1,10 @@
 """Launcher file for trpo sawyer reacher training."""
 
-from garage.algos import TRPO
 from garage.baselines import LinearFeatureBaseline
 from garage.envs.mujoco.sawyer import ReacherEnv
 from garage.misc.instrument import run_experiment
 from garage.policies import GaussianMLPPolicy
+from garage.theano.algos import TRPO
 from garage.theano.envs import TheanoEnv
