+++ garage/algos/__init__.py
@@ -4,12 +4,4 @@ from garage.algos.batch_polopt import BatchPolopt
 from garage.algos.batch_polopt import BatchSampler
 from garage.algos.cem import CEM
 from garage.algos.cma_es import CMAES
-from garage.algos.ddpg import DDPG
-from garage.algos.vpg import VPG
-from garage.algos.erwr import ERWR  # noqa: I100
 from garage.algos.nop import NOP
-from garage.algos.npo import NPO
-from garage.algos.ppo import PPO
-from garage.algos.reps import REPS
-from garage.algos.tnpg import TNPG
-from garage.algos.trpo import TRPO
