+++ tests/envs/test_sawyer_envs.py
@@ -4,7 +4,6 @@ import unittest
 
 import numpy as np
 
-from garage.algos import TRPO
 from garage.baselines import LinearFeatureBaseline
 from garage.envs.mujoco.sawyer import BinSortingEnv
 from garage.envs.mujoco.sawyer import BlockStackingEnv
@@ -12,6 +11,7 @@ from garage.envs.mujoco.sawyer import PickAndPlaceEnv
 from garage.envs.mujoco.sawyer import ReacherEnv
 from garage.envs.util import spec
 from garage.policies import GaussianMLPPolicy
+from garage.theano.algos import TRPO
 
 
 def run_bin_sorting(*_):
