+++ tests/test_adaptive_std.py
@@ -1,9 +1,9 @@
 from nose2.tools import such
 
-from garage.algos import TRPO
 from garage.baselines import ZeroBaseline
 from garage.envs.box2d import CartpoleEnv
 from garage.policies import GaussianMLPPolicy
+from garage.theano.algos import TRPO
 from garage.theano.envs import TheanoEnv
 
 with such.A("adaptive_std") as it:
